requirements.txt
args==0.1.0
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
clint==0.5.1
idna==3.4
pycodestyle==2.10.0
python-decouple==3.6
requests==2.28.1
tomli==2.0.1
urllib3==1.26.12
